# swagger.yaml
openapi: 3.0.0
paths:
...
  /api/v1/contents:
    get:
      operationId: ContentsController_find
      summary: Get contents paginated list
      parameters:
        - name: filters
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                filters[id][$eq]: 1
          schema:
            type: object
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            type: integer
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            type: integer
        - name: sorts
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                sorts[id]: asc
          schema:
            type: object
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - article
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Content'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: &ref_4
        - Contents
      security:
        - bearer: []
    post:
      operationId: ContentsController_create
      summary: Create new content
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateContentDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: *ref_4
      security:
        - bearer: []
  /api/v1/contents/all:
    get:
      operationId: ContentsController_findAll
      summary: Get all contents
      parameters:
        - name: filters
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                filters[id][$eq]: 1
          schema:
            type: object
        - name: sorts
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                sorts[id]: asc
          schema:
            type: object
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - article
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Content'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: *ref_4
      security:
        - bearer: []
  /api/v1/contents/{id}:
    get:
      operationId: ContentsController_findOne
      summary: Get specific content
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - article
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_4
      security:
        - bearer: []
    put:
      operationId: ContentsController_update
      summary: Update specific content
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateContentDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_4
      security:
        - bearer: []
    delete:
      operationId: ContentsController_remove
      summary: Delete specific content
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_4
      security:
        - bearer: []
  /api/v1/tags:
    get:
      operationId: TagsController_find
      summary: Get tags paginated list
      parameters:
        - name: filters
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                filters[id][$eq]: 1
          schema:
            type: object
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            type: integer
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            type: integer
        - name: sorts
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                sorts[id]: asc
          schema:
            type: object
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - articles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: &ref_5
        - Tags
      security:
        - bearer: []
    post:
      operationId: TagsController_create
      summary: Create new tag
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateTagDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: *ref_5
      security:
        - bearer: []
  /api/v1/tags/all:
    get:
      operationId: TagsController_findAll
      summary: Get all tags
      parameters:
        - name: filters
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                filters[id][$eq]: 1
          schema:
            type: object
        - name: sorts
          required: false
          in: query
          examples:
            Example 1:
              value: {}
            Example 2:
              value:
                sorts[id]: asc
          schema:
            type: object
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - articles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
      tags: *ref_5
      security:
        - bearer: []
  /api/v1/tags/{id}:
    get:
      operationId: TagsController_findOne
      summary: Get specific tag
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
        - name: includes[]
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - articles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_5
      security:
        - bearer: []
    put:
      operationId: TagsController_update
      summary: Update specific tag
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateTagDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_5
      security:
        - bearer: []
    delete:
      operationId: TagsController_remove
      summary: Delete specific tag
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_5
      security:
        - bearer: []
...
info:
  title: Nest Articles App
  description: RESTful API with NestJS and TypeORM
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3333
    description: Dev Server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        rowCount:
          type: integer
          minimum: 0
        pageCount:
          type: integer
          minimum: 0
      required:
        - page
        - pageSize
        - rowCount
        - pageCount
...
    Content:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
        article_id:
          type: integer
          minimum: 1
        body:
          type: string
      required:
        - id
        - created_at
        - updated_at
        - article_id
        - body
    CreateOrUpdateContentDto:
      type: object
      properties:
        article_id:
          type: integer
          minimum: 1
        body:
          type: string
      required:
        - article_id
        - body
    Tag:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
        name:
          type: string
          maxLength: 100
      required:
        - id
        - created_at
        - updated_at
        - name
    CreateOrUpdateTagDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
      required:
        - name
...